requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.71
botocore==1.27.71
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Django==4.1.1
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
future==0.18.2
GDAL @ file:///C:/Users/chinni/Downloads/GDAL-3.4.3-cp38-cp38-win_amd64.whl
geocoder==1.38.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.2.1
ratelim==0.1.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
